refactor: split tracedecay.rs lifecycle and locking into submodules
